% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFiles.R
\docType{data}
\name{example.bed}
\alias{example.bed}
\title{Example Gene Annotation in Bed-Format}
\format{A file with three column Chr, Start and End.}
\description{
This file contains some example lines to represent a typical bed file that can be used
to try the corresponding functions.
}
\details{
The file is locate din the /extdata folder of the package and is accessible after installation via
\code{system.file("extdata","example.bed", package="GenomicTools.fileHandler")}
}
\author{
Daniel Fischer
}
\keyword{data}
