% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFiles.R
\docType{data}
\name{example.fastq}
\alias{example.fastq}
\title{Example Sequencing Reads in fastq-Format}
\usage{
system.file("extdata","example.fastq", package="GenomicTools.fileHandler")
}
\description{
This file contains some example reads to represent a typical fastq file that can be used
to try the corresponding functions.
}
\details{
The file is locate din the /extdata folder of the package and is accessible after installation via
\code{system.file("extdata","example.fastq", package="GenomicTools.fileHandler")}
}
\author{
Daniel Fischer
}
\keyword{data}
