% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFiles.R
\docType{data}
\name{example.gff}
\alias{example.gff}
\title{Example Gene Annotation in gff-Format}
\usage{
system.file("extdata","example.gff", package="GenomicTools.fileHandler")
}
\description{
This file contains some example gene annotations to represent a typical gff file that can be used
to try the corresponding functions.
}
\details{
The file is locate din the /extdata folder of the package and is accessible after installation via
\code{system.file("extdata","example.gff", package="GenomicTools.fileHandler")}
}
\author{
Daniel Fischer
}
\keyword{data}
