% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFiles.R
\docType{data}
\name{example.vcf}
\alias{example.vcf}
\title{Example Variant data in vcf-Format}
\usage{
system.file("extdata","example.vcf", package="GenomicTools.fileHandler")
}
\description{
This file contains some example variants to represent a typical vcf file that can be used
to try the corresponding functions.
}
\details{
The file is locate din the /extdata folder of the package and is accessible after installation via
\code{system.file("extdata","example.vcf", package="GenomicTools.fileHandler")}
}
\author{
Daniel Fischer
}
\keyword{data}
