% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFiles.R
\docType{data}
\name{example2.gtf.gz}
\alias{example2.gtf.gz}
\title{Example Gene Annotation in zipped gtf-Format}
\usage{
system.file("extdata","example2.gtf.gz", package="GenomicTools.fileHandler")
}
\description{
This file contains some example gene annotations to represent a typical zipped gtf file that can be used
to try the corresponding functions.
}
\details{
The file is locate din the /extdata folder of the package and is accessible after installation via
\code{system.file("extdata","example2.gtf.gz", package="GenomicTools.fileHandler")}
}
\author{
Daniel Fischer
}
\keyword{data}
