% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFA.R
\name{exportFA}
\alias{exportFA}
\title{Exporting a Fasta File.}
\usage{
exportFA(fa, file = NULL)
}
\arguments{
\item{fa}{fasta object}

\item{file}{Character, specifies filename/path}
}
\value{
A fasta file
}
\description{
This function exports a standard fasta file.
}
\details{
This function exports a fasta object to a standard fasta file. If no file name is given, the variable name will be used instead.
}
\examples{
\dontrun{
novelBed <- data.frame(Chr=c(11,18,3),
Start=c(72554673, 62550696, 18148822),
End=c(72555273, 62551296, 18149422),
Gene=c("LOC1", "LOC2", "LOC3"))

myFasta <- getFastaFromBed(novelBed, species="Bos taurus", 
+                          fastaFolder="/home/daniel/fasta/")

exportFA(myFasta, file="myFasta.fa")
}
}
\author{
Daniel Fischer
}
