% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportGTF.R
\name{exportGTF}
\alias{exportGTF}
\title{Exporting a GTF File.}
\usage{
exportGTF(x, file)
}
\arguments{
\item{x}{gtf-object}

\item{file}{Character, specifies filename/path}
}
\value{
A gtf file
}
\description{
This function exports a standard gtf file.
}
\details{
This function exports a gtf-object to a standard gtf file.
}
\author{
Daniel Fischer
}
