\name{GenomicTools.fileHandler-package}
\alias{R/GenomicTools.fileHandler-package}
\docType{package}

\title{
  Loose Collection of I/O Functions Needed in Genomic Data Analysis.
}

\description{
  GenomicTools.fileHandler is a loose collection of I/O Functions Needed in Genomic Data Analysis.
}

\details{
  \tabular{ll}{
    Package: \tab GenomicTools.fileHandler\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2018-07-26\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@luke.fi>
}

\keyword{multivariate}

