% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTools.fileHandler.R
\docType{package}
\name{GenomicTools.fileHandler-package}
\alias{GenomicTools.fileHandler}
\alias{GenomicTools.fileHandler-package}
\title{R Package To Handle Files From Genomic Data
GenomicTools.fileHandler is a loose collection of I/O Functions Needed in Genomic Data Analysis}
\description{
\tabular{ll}{ Package: \tab GenomicTools.fileHandler\cr Type: \tab Package\cr Version: \tab
0.1.5.23\cr Date: \tab 2022-01-23\cr License: \tab GPL\cr LazyLoad: \tab yes\cr
}
}
\author{
Daniel Fischer

Maintainer: Daniel Fischer <daniel.fischer@luke.fi>
}
