% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBlastTab.R
\name{importBlastTab}
\alias{importBlastTab}
\title{Import a Tab Delimited Blast Output File}
\usage{
importBlastTab(file)
}
\arguments{
\item{file}{Filename}
}
\value{
A data.frame
}
\description{
This function imports a tab delimited blast output.
}
\details{
This function imports a tab delimited blast output file, currently the same as \code{read.table}
}
\author{
Daniel Fischer
}
