% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFA.R
\name{importFA}
\alias{importFA}
\title{Importing a Fasta File.}
\usage{
importFA(file)
}
\arguments{
\item{file}{Specifies the filename/path}
}
\value{
An object of class \code{fa} containing the sequences. The names correspond to the sequence names given in the fasta file.
}
\description{
This function imports a standard fasta file
}
\details{
This function imports a standard fasta file. It assumes that label and sequence lines are alternating,
meaning in the odd lines are the sequence names given, starting with > and in the even rows are the 
corresponding sequences.
}
\examples{

 \dontrun{
   importFA(file="myFasta.fa")
 }
}
\seealso{
print.fa, summary.fa
}
\author{
Daniel Fischer
}
