% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFQ.R
\name{importFQ}
\alias{importFQ}
\title{Importing a Fastq File.}
\usage{
importFQ(file)
}
\arguments{
\item{file}{Specifies the filename/path}
}
\value{
An object of class \code{fq} containing the sequences and the quality meausure. The names correspond to the sequence names given in the fasta file.
}
\description{
This function imports a standard fastq file
}
\details{
This function imports a standard fastq file that consists out of blocks of four lines per entry
}
\examples{

 # Define here the location on HDD for the example file
   fpath <- system.file("extdata","example.fastq", package="GenomicTools.fileHandler")
 # Import the example fastq file  
   fastqFile <- importFQ(file=fpath)
 
}
\seealso{
print.fq, summary.fq
}
\author{
Daniel Fischer
}
