% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGFF3.R
\name{importGFF3}
\alias{importGFF3}
\title{importGFF3}
\usage{
importGFF3(gff, chromosomes)
}
\arguments{
\item{gff}{file or folder}

\item{chromosome}{The chromosome to import}
}
\value{
A gff object
}
\description{
Import a GFF3 file
}
\details{
This function imports a standard gff3 file.
}
\author{
Daniel Fischer
}
