% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPED.R
\name{importPED}
\alias{importPED}
\title{importPED}
\usage{
importPED(file, n, snps, which, split = "\\t| +", sep = ".",
  na.strings = "0", lex.order = FALSE, verbose = TRUE)
}
\arguments{
\item{file}{ped filename}

\item{n}{Number of samples to read}

\item{snps}{map filename}

\item{which}{Names of SNPS to import}

\item{split}{Columns separator in ped file}

\item{sep}{Character that separates Alleles}

\item{lex.order}{Logical, lexicographical order}

\item{verbose}{Logical, verbose output}

\item{ns.strings}{Definition for missing values}
}
\value{
a pedmap object
}
\description{
Import a PED/MAP file pair
}
\details{
This function is to a large extend taken from \code{snpStat::read.pedmap}, but here is internally the \code{data.table::fread} function used
that resulted in much faster file processing.
}
\author{
Daniel Fischer
}
