% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSTARLog.R
\name{importSTARLog}
\alias{importSTARLog}
\title{importSTARLog}
\usage{
importSTARLog(file)
}
\arguments{
\item{file}{The file name}
}
\value{
a data frame
}
\description{
Import the Log-File from STAR
}
\details{
This function imports the Log file from STAR
}
\author{
Daniel Fischer
}
