% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importXML.R
\name{importXML}
\alias{importXML}
\title{importXML}
\usage{
importXML(folder, seqNames = NULL, which = NULL, idTH = 0.8,
  verbose = TRUE)
}
\arguments{
\item{folder}{Character, folder path}

\item{seqNames}{Names of sequences}

\item{which}{Which sequences to import}

\item{idTH}{Use the threshold as cut-off}

\item{verbose}{Logical, verbose output}
}
\value{
An XML object
}
\description{
Import an Blast XML file
}
\details{
This function imports XML files as provided as Blast output, it is mainly aimied to import the output of the hoardeR package
}
\author{
Daniel Fischer
}
