% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalReads.R
\name{plotTotalReads}
\alias{plotTotalReads}
\title{plotTotalReads}
\usage{
plotTotalReads(STARLog)
}
\arguments{
\item{STARLog}{A STARLog object}
}
\value{
A plot
}
\description{
Plot the total reads
}
\details{
This function plots the total reads from a STARlog object

Part of the diagnostic plot series for of the STARLog. The function accepts also a list of STARLogs and creates then comparative boxplots
}
\author{
Daniel Fischer
}
