% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUniquelyMappedReads.R
\name{plotUniquelyMappedReads}
\alias{plotUniquelyMappedReads}
\title{plotUniquelyMappedReads}
\usage{
plotUniquelyMappedReads(STARLog)
}
\arguments{
\item{STARLog}{A STARLog object}
}
\value{
A plot
}
\description{
Plot the uniquely mapped reads
}
\details{
This function plots the percenage of uniquely reads from a STARlog object

Part of the diagnostic plot series for of the STARLog. The function accepts also a list of STARLogs and creates then comparative boxplots
}
\author{
Daniel Fischer
}
