% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.bed}
\alias{print.bed}
\title{Print a bed Object}
\usage{
\method{print}{bed}(x, n = 6, ...)
}
\arguments{
\item{x}{Object of class \code{bed}.}

\item{...}{Additional parameters}
}
\description{
Prints a \code{bed} object.
}
\details{
The print function displays a bed object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
