% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.fa}
\alias{print.fa}
\title{Print a fa Object}
\usage{
\method{print}{fa}(x, n = 2, seq.out = 50, ...)
}
\arguments{
\item{x}{Object of class \code{fa}.}

\item{n}{Number of sequences to display}

\item{seq.out}{Length of the subsequence to display}

\item{...}{Additional parameters}
}
\description{
Prints a \code{fa} object.
}
\details{
The print function displays a fa object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
