% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.featureCounts}
\alias{print.featureCounts}
\title{Print a featureCounts Object}
\usage{
\method{print}{featureCounts}(x, ...)
}
\arguments{
\item{x}{Object of class \code{featureCounts}.}

\item{...}{Additional parameters}
}
\description{
Prints an \code{featureCounts} object.
}
\details{
The print function displays a featureCounts object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
