% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.fq}
\alias{print.fq}
\title{Print a fq Object}
\usage{
\method{print}{fq}(x, n = 2, seq.out = 50, print.qual = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{fq}.}

\item{n}{Number of sequences to display}

\item{seq.out}{Length of the subsequence to display}

\item{print.qual}{Logical, shall the quality measures also be printed}

\item{...}{Additional parameters}
}
\description{
Prints a \code{fq} object.
}
\details{
The print function displays a fa object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
