% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.gtf}
\alias{print.gtf}
\title{Print a gtf Object}
\usage{
\method{print}{gtf}(x, n = 6, ...)
}
\arguments{
\item{x}{Object of class \code{gtf}.}

\item{n}{Number of lines to print}

\item{...}{Additional parameters}
}
\description{
Prints a \code{gtf} object.
}
\details{
The print function displays a bed object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
