% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.pedMap}
\alias{print.pedMap}
\title{Print a pedMap Object}
\usage{
\method{print}{pedMap}(x, n = 6, m = 6, ...)
}
\arguments{
\item{x}{Object of class \code{pedMap}.}

\item{n}{Number of samples to display}

\item{m}{Number of columns to display}

\item{...}{Additional parameters}
}
\description{
Prints an \code{pedMap} object.
}
\details{
The print function displays a pedMap object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
