% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3print.R
\docType{methods}
\name{print.vcf}
\alias{print.vcf}
\title{Print a vcf Object}
\usage{
\method{print}{vcf}(x, n = 6, m = 6, fullHeader = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{vcf}.}

\item{n}{Number of samples to display}

\item{m}{Number of columns to display}

\item{fullHeader}{Logical, shall the whole header be printed}

\item{...}{Additional parameters}
}
\description{
Prints an \code{vcf} object.
}
\details{
The print function displays a vcf object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}
