% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3summary.R
\docType{methods}
\name{summary.bed}
\alias{summary.bed}
\title{Summary of a bed Object}
\usage{
\method{summary}{bed}(object, ...)
}
\arguments{
\item{...}{Additional parameters}

\item{x}{Object of class \code{bed}.}
}
\description{
Summarizes a \code{bed} object.
}
\details{
The summary function displays an informative summary of a bed object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{summary}
