% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3summary.R
\docType{methods}
\name{summary.fa}
\alias{summary.fa}
\title{Summary of a fa Object}
\usage{
\method{summary}{fa}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bed}.}

\item{...}{Additional parameters}
}
\description{
Summarizes a \code{fa} object.
}
\details{
The summary function displays an informative summary of a fa object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{summary}
