% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3summary.R
\docType{methods}
\name{summary.featureCounts}
\alias{summary.featureCounts}
\title{Summary of a featureCounts Object}
\usage{
\method{summary}{featureCounts}(object, ...)
}
\arguments{
\item{object}{Object of class \code{featureCounts}.}

\item{...}{Additional parameters}
}
\description{
Summarizes a \code{featureCounts} object.
}
\details{
The summary function displays an informative summary of a featureCounts object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{summary}
