% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3summary.R
\docType{methods}
\name{summary.fq}
\alias{summary.fq}
\title{Summary of a fq Object}
\usage{
\method{summary}{fq}(object, ...)
}
\arguments{
\item{object}{Object of class \code{fq}.}

\item{...}{Additional parameters}
}
\description{
Summarizes a \code{fq} object.
}
\details{
The summary function displays an informative summary of a fq object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{summary}
