% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3summary.R
\docType{methods}
\name{summary.gtf}
\alias{summary.gtf}
\title{Summary of a gtf Object}
\usage{
\method{summary}{gtf}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gtf}.}

\item{...}{Additional parameters}
}
\description{
Summarizes a \code{gtf} object.
}
\details{
The summary function displays an informative summary of a gtf object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{summary}
