% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3summary.STARLog.R
\docType{methods}
\name{summary.STARLog}
\alias{summary.STARLog}
\title{Summary of a STARLog Object}
\usage{
\method{summary}{STARLog}(object, ...)
}
\arguments{
\item{object}{Object of class \code{STARLog}.}

\item{...}{Additional parameters}
}
\description{
Summarizes a \code{STARLog} object.
}
\details{
The summary function displays an informative summary of a STARLog object
}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{summary}
