% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVCF.R
\name{exportVCF}
\alias{exportVCF}
\title{Exporting a VCF object.}
\usage{
exportVCF(vcf, file = NULL)
}
\arguments{
\item{vcf}{vcf object}

\item{file}{Character, specifies filename/path}
}
\value{
A vcf file
}
\description{
This function exports a vcf object into a standard vcf file.
}
\details{
This function exports a vcf object to a standard vcf file. If no file name is given, the variable name will be used instead.
}
\examples{
 # Define here the location on HDD for the example file
   fpath <- system.file("extdata","example.vcf", package="GenomicTools.fileHandler")
 # Import the example fasta file  
 #  vcfFile <- importVCF(file=fpath)
   
 # myfile <- file.path(tempdir(), "myLocs.vcf")
   
  exportVCF(newVCF, file=myfile)


}
\author{
Daniel Fischer
}
