% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBed.R
\name{importBed}
\alias{importBed}
\title{Importing a Bed File.}
\usage{
importBed(file, header = FALSE)
}
\arguments{
\item{file}{Specifies the filename/path}

\item{header}{Logical, is a header present}
}
\value{
A \code{data.frame}
}
\description{
This function imports a standard bed file
}
\details{
This function imports a standard bed-file into a data.frame. It is basically a convenience wrapper around \code{read.table}
}
\examples{
\dontrun{
   novelBed <- data.frame(Chr=c(11,18,3),
                          Start=c(72554673, 62550696, 18148822),
                          End=c(72555273, 62551296, 18149422),
                          Gene=c("LOC1", "LOC2", "LOC3"))
   
   exportBed(novelBed, file="myLocs.bed")
   
   novelBed.imp <- importBed(file="myLocs.bed")
 }
 
 @export
}
\seealso{
[exportBed], [read.table]
}
\author{
Daniel Fischer
}
