% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSTARLog.R
\name{importSTARLog}
\alias{importSTARLog}
\title{importSTARLog}
\usage{
importSTARLog(dir, recursive = TRUE, log = FALSE, finalLog = TRUE,
  verbose = TRUE)
}
\arguments{
\item{dir}{The directory name}

\item{recursive}{Logical, check for sub-directories}

\item{log}{boolean, import also log file}

\item{finalLog}{boolean, import also final_log file}

\item{verbose}{Logical, talkactive function feedback}
}
\value{
a data frame
}
\description{
Import the Log-File from STAR
}
\details{
This function imports the Log file from STAR
}
\author{
Daniel Fischer
}
