% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVCF.R
\name{importVCF}
\alias{importVCF}
\title{importVCF}
\usage{
importVCF(file, na.seq = "./.")
}
\arguments{
\item{file}{The file name}

\item{na.seq}{The missing value definition}
}
\value{
A vcf object
}
\description{
Import a VCF function
}
\details{
This function imports a VCF file.

In case the logicl flag 'phased' is set to TRUE then the genotypes are expected to be in the format 0|0, otherwise they are expected
to be like 0/1 .

The example file was downloaded from here:

ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/pilot_data/release/2010_07/exon/snps/
}
\examples{

 # Define here the location on HDD for the example file
   fpath <- system.file("extdata","example.vcf", package="GenomicTools.fileHandler")
 # Import the example vcf file  
   importVCF(fpath)

}
\author{
Daniel Fischer
}
