% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVCF.R
\name{importVCF}
\alias{importVCF}
\title{importVCF}
\usage{
importVCF(file, na.seq = "./.")
}
\arguments{
\item{file}{The file name}

\item{na.seq}{The missing value definition}
}
\value{
A vcf object
}
\description{
Import a VCF function
}
\details{
This function imports a VCF file
}
\author{
Daniel Fischer
}
